added function to post when string is bigger than 220 char
Added a function to post the tweet even when the string gets bigger than 220 characters, posting the remaining of the string as a response to the first part. Also added a bit of code to store the menu data from the whole week into a csv for future data analysis.